requirements.txt
gradientai
python-dotenv
Flask
Flask-Cors
jsonify
fastapi
pydantic
uvicorn
